+++ camel/environments/__init__.py
@@ -11,6 +11,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
-from .base import BaseEnvironment
+from .base import BaseEnvironment, Response
 
-__all__ = ["BaseEnvironment"]
+__all__ = ["BaseEnvironment", "Response"]
