+++ camel/environments/base.py
@@ -13,18 +13,82 @@
 # ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
 
 from abc import ABC, abstractmethod
+from datetime import datetime
 from typing import Any, Dict, List, Optional, Tuple
 
 from pydantic import BaseModel, Field
 
 from camel.agents import ChatAgent
+from camel.configs import BaseConfig
 from camel.datasets.base import BaseDataset
 from camel.extractors.base import BaseExtractor
 from camel.logger import get_logger
-from camel.verifiers.base import BaseVerifier, VerificationResult
-from camel.verifiers.models import Response, TaskType, VerificationStatus
+from camel.types import ModelType
+from camel.verifiers.base import (
+    BaseVerifier,
+    VerificationResult,
+)
+from camel.verifiers.models import (
+    MachineInfo,
+    TaskType,
+    VerificationStatus,
+    VerifierConfig,
+)
 
 logger = get_logger(__name__)
+
+
+class Response(BaseModel):
+    r"""Response schema for LLM generations with metadata and verification
+    info.
+
+    Contains the input context, generated output, and metadata for
+    verification.
+
+    Attributes:
+        problem_id: Unique identifier for the problem
+        source: Source of the problem/task
+        task_type: Type of task being performed
+        final_answer: Reference solution if available
+        verification_info: Additional info needed for verification
+        metadata: General purpose metadata about the generation
+        prompt: Input prompt for the LLM
+        llm_response: Generated response from the LLM
+        model_type: Type of the LLM used
+        generation_config: Configuration used for generation
+        machine_info: Information about the execution environment
+        timestamp: When the response was generated (UTC)
+    """
+
+    problem_id: str = Field(description="Unique identifier for the problem")
+    source: str = Field(description="Source of the problem/task")
+    task_type: TaskType = Field(description="Type of task being performed")
+    final_answer: Optional[str] = Field(
+        None, description="Reference solution if available"
+    )
+    verification_info: Dict[str, Any] = Field(
+        default_factory=dict, description="Information needed for verification"
+    )
+    verification_config: Optional[VerifierConfig] = Field(
+        None, description="Configuration for verification behavior"
+    )
+    metadata: Dict[str, Any] = Field(
+        default_factory=dict,
+        description="Additional metadata about the generation",
+    )
+    prompt: str = Field(description="Input prompt for the LLM")
+    llm_response: str = Field(description="Generated response from the LLM")
+    model_type: ModelType = Field(description="Type of the LLM used")
+    generation_config: BaseConfig = Field(
+        description="Generation parameters used"
+    )
+    machine_info: Optional[MachineInfo] = Field(
+        None, description="Execution environment info"
+    )
+    timestamp: datetime = Field(
+        default_factory=datetime.utcnow,
+        description="When the response was generated (UTC)",
+    )
 
 
 class Observation(BaseModel):
